IS-1821: Kjøre cronJob om natta
