+++ src/main/kotlin/no/nav/syfo/cronjob/CronjobModule.kt
@@ -57,7 +57,7 @@ fun cronjobModule(
         dialogmeldingStatusCronjob,
         verifyPartnerIdCronjob,
         suspensjonCronjob,
-        // verifyBehandlereForKontorCronjob,
+        verifyBehandlereForKontorCronjob,
     ).forEach {
         launchBackgroundTask(
             applicationState = applicationState,
